Explicitly set editing to false
Previously, when deleting the last account and then immediately adding a
new one, the accounts view controller and its belonging table view would
be in an editing state when appearing after the new account had been
added. This was confusing since the edit button was hidden after
deleting the last account, and the empty state label had been shown.

To fix this, this commit explicitly sets editing to false after deleting
the last account.