Persist account data in the Keychain only
Previously, account data was split across CoreData and the iOS Keychain
with the everything except the secret stored in CoreData.

While this allowed me to use conveniences like
`NSFetchedResultsController` it also made the code more complex.

When restoring from an unencrypted backup only the data stored in
CoreData would be restored causing the app to crash.

This commit stores all the account data in the Keychain by serializing
it using a custom versioned binary format.