Persist counters incremented from search results
Previously, if the user searched for an account with a counter based
password and clicked the next button, the value for the account would
update but the new counter value would not be persisted to the keychain
since the cell had no delegate assigned.

When transitioning back to the accounts view after incrementing the
counter, the account value shown there would also be stale.

This commit fixes both these issues.