Stop animating continuously
Doing this created a bunch of edge cases where animation would stop for
some reason. One example of this was when the view in question
disappeared off screen. When it reappeared it would not start animating
until the next period of the password.

Having something update visible cells every second is both more reliable
and performant. The progress view updates are still animated, which
provides a good experience for users while still being very manageable.