Use 64-bit integers
On the iPhone 4S, which uses the 32-bit Apple A5 processor, password
values would be incorrectly generated. On the iPhone 5S, which uses the
64-bit Apple A7 processor, and above they would not.

The reason for the incorrectly generated values was the use of the `Int`
type, which on 32-bit systems is the same as `Int32`.

This commit switches to explicitly use 64-bit integers using the `Int64`
type.