Improve no camera access view
Previously, when disallowing Tofu to access the camera, a black view
would be shown instead of the message about camera access. This commit
fixes this.

This commit also ensures the prompt to scan a QR code isn't shown unless
the camera is active.

It also improves the actual message shown.