Prevent crashes when trying to use too many digits
As far as I understand, the HOTP algorithm as described by [RFC4226]
supports a maximum of 9 digits. Trying to use more (by scanning a QR
code of a URL with a larger digits paramater) previously caused the app
to crash. To avoid this, this commit prevents trying to generare OTPs
with more than 9 digits.

RFC4226 also states that "Implementations MUST extract a 6-digit code at
a minimum [...]". Thererfore, this commit also prevents trying to
generate OTPs with fewer than 6 digits.

[RFC4226]: https://tools.ietf.org/html/rfc4226